test(app-core): colocate preAction hook test so it actually loads
The suite sat in __tests__/ but imported ./preaction.ts and mocked
../argv, ../banner and ../../utils/globals as if it were colocated, so
the module never resolved and the real collaborators were used. Move it
next to preaction.ts, where every specifier resolves, and extend the
cases to cover process-title, verbose and ELIZA_HIDE_BANNER branches.
